create a separate create notebook launcher
